# backend/oas/user/openapi.yaml
openapi: 3.0.1
info:
  title: OQTOPUS Cloud User API
  version: '1.0'
  contact:
    name: oqtopus-team
    email: oqtopus-team[at]googlegroups.com
  description: OQTOPUS Cloud User API. This API is used to interact with the OQTOPUS Cloud service. The API provides endpoints to manage devices, tasks, and results.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local server url
paths:
  /devices:
    get:
      tags:
        - device
      summary: List available devices
      description: List available devices
      operationId: listDevices
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Returns a list of available devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/devices.DeviceInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.UnauthorizedError'
              example:
                detail: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.InternalServerError'
  /devices/{deviceId}:
    get:
      tags:
        - device
      summary: Get specified device details
      description: get device
      operationId: getDevice
      security:
        - BearerAuth: []
      parameters:
        - name: deviceId
          in: path
          description: Device identifier
          required: true
          schema:
            type: string
          example: Kawasaki
      responses:
        '200':
          description: task response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/devices.DeviceInfo'
        '404':
          description: Device with deviceId not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.NotFoundError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.InternalServerError'
...
  /results/sampling/{taskId}:
    get:
      tags:
        - result
      summary: Get result of a sampling task
      description: Gets a result of quantum task.<br/><br/> Any finished quantum task must generate a result. The result contains either simulation result (for successfully completed tasks) or reason why simulation outcome is not provided (for unsuccessful and cancelled tasks)<br/><br/>Operation is valid only for COMPLETED, FAILED and CANCELLED tasks
      operationId: getSamplingResult
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '200':
          description: Return quantum task result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/results.SamplingResultDef'
              examples:
                success sampling:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: SUCCESS
                    result:
                      '11': 4980
                      '00': 5020
                    transpiledCode: OPENQASM 3;include "stdgates.inc";qubit[2] _all_qubits;let q = _all_qubits[0:1];h q[0];cx q[0], q[1];
                failure:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: FAILURE
                    reason: Invalid device parameter in request
                cancel:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: CANCELLED
                    reason: User cancelled
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.BadRequest'
              example:
                detail: Bad request malformed input data
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.UnauthorizedError'
              example:
                detail: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.NotFoundError'
              example:
                detail: task not found
  /results/estimation/{taskId}:
    get:
      tags:
        - result
      summary: Get result of estimation task
      description: Gets a result of quantum task.<br/><br/> Any finished quantum task must generate a result. The result contains either simulation result (for successfully completed tasks) or reason why simulation outcome is not provided (for unsuccessful and cancelled tasks)<br/><br/>Operation is valid only for COMPLETED, FAILED and CANCELLED tasks
      operationId: getEstimationResult
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          description: Task identifier
          schema:
            type: string
      responses:
        '200':
          description: Return quantum task result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/results.EstimationResultDef'
              examples:
                success estimation:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: SUCCESS
                    result:
                      - 1.5
                      - 2.8
                    transpiledCode: OPENQASM 3;include "stdgates.inc";qubit[2] _all_qubits;let q = _all_qubits[0:1];h q[0];cx q[0], q[1];
                failure:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: FAILURE
                    reason: Invalid device parameter in request
                cancel:
                  value:
                    taskId: 7af020f6-2e38-4d70-8cf0-4349650ea08c
                    status: CANCELLED
                    reason: User cancelled
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.BadRequest'
              example:
                detail: Bad request malformed input data
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.UnauthorizedError'
              example:
                detail: Unauthorized
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error.NotFoundError'
              example:
                detail: task not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    devices.CalibrationData:
      description: Calibration data available only for 'QPU' devices
      type: object
      properties:
        qubitConnectivity:
          type: array
          items:
            type: string
            format: (number, number)
          example:
            - (1,4)
            - (4,5)
            - (5,8)
        t1:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            '0': 55.51
            '1': 37.03
            '2': 57.13
        t2:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            '0': 99.31
            '1': 111.03
            '2': 30.12
        roError:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            '0': 0.0467
            '1': 0.18
            '2': 0.355
        gateError:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: number
              format: float
          example:
            sx:
              '0': 0.00692
              '1': 0.00296
              '2': 0.072
        measProb0As1:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            '0': 0.00608
            '1': 0.012
            '2': 0.248
        measProb1As0:
          type: object
          additionalProperties:
            type: number
            format: float
          example:
            '0': 0.0281
            '1': 0.386
            '2': 0.0811
        gateDuration:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: number
              format: float
          example:
            sx:
              '0': 29.3
              '1': 50.9
              '2': 45.4
    devices.DeviceInfo:
      type: object
      properties:
        deviceId:
          type: string
          example: SVSim
        deviceType:
          type: string
          enum:
            - QPU
            - simulator
          example: simulator
        status:
          type: string
          enum:
            - AVAILABLE
            - NOT_AVAILABLE
          example: AVAILABLE
        restartAt:
          description: Parameter mandatory and valid for 'NOT_AVAILABLE' devices
          type: string
          format: date-time
          example: '2022-10-19 11:45:34'
        nPendingTasks:
          type: integer
          example: 8
        nQubits:
          type: integer
          example: 39
        nNodes:
          description: Parameter valid only for 'simulator' devices
          type: integer
          example: 512
        basisGates:
          type: array
          items:
            type: string
          example:
            - x
            - 'y'
            - z
            - h
            - s
            - sdg
            - t
            - tdg
            - rx
            - ry
            - rz
            - cx
            - cz
            - swap
            - u1
            - u2
            - u3
            - u
            - p
            - id
            - sx
            - sxdg
        supportedInstructions:
          type: array
          items:
            type: string
          example:
            - measure
            - barrier
            - reset
        calibrationData:
          $ref: '#/components/schemas/devices.CalibrationData'
        calibratedAt:
          description: Parameter available only for 'QPU' devices with available calibration data
          type: string
          format: date-time
          example: '2022-10-19 11:45:34'
        description:
          type: string
          example: State vector-based quantum circuit simulator
      required:
        - deviceId
        - deviceType
        - status
        - nPendingTasks
        - basisGates
        - supportedInstructions
        - description
      example:
        - deviceId: SVSim
          deviceType: simulator
          status: AVAILABLE
          nPendingTasks: 8
          nQubits: 39
          nNodes: 512
          basisGates:
            - x
            - 'y'
            - z
            - h
            - s
            - sdg
            - t
            - tdg
            - rx
            - ry
            - rz
            - cx
            - cz
            - swap
            - u1
            - u2
            - u3
            - u
            - p
            - id
            - sx
            - sxdg
          supportedInstructions:
            - measure
            - barrier
            - reset
          description: State vector-based quantum circuit simulator
    error.UnauthorizedError:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
    error.InternalServerError:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
    error.NotFoundError:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
    tasks.TaskId:
      type: string
      format: uuid
      example: 7af020f6-2e38-4d70-8cf0-4349650ea08c
...
    error.BadRequest:
      type: object
      properties:
        detail:
          type: string
      required:
        - detail
...
    results.ResultStatus:
      type: string
      enum:
        - SUCCESS
        - FAILURE
        - CANCELLED
      default: SUCCESS
      example: SUCCESS
    results.SamplingResultDef:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/tasks.TaskId'
        status:
          $ref: '#/components/schemas/results.ResultStatus'
        result:
          description: Provided only for successful results
          type: object
          additionalProperties:
            type: integer
            minProperties: 2
          example:
            '11': 4980
            '00': 5020
        reason:
          description: Provided only for unsuccessful (failed, cancelled) results
          type: string
        transpiledCode:
          description: Provided only for tasks with transpilation (tasks created with skipTranspilation=False), if traspilation was successfull
          type: string
        qubitAllocation:
          description: Parameter valid only for QPU devices
          type: object
          additionalProperties:
            type: integer
          example:
            '0': 0
            '1': 4
      required:
        - taskId
        - status
    results.EstimationResultDef:
      type: object
      properties:
        taskId:
          $ref: '#/components/schemas/tasks.TaskId'
        status:
          $ref: '#/components/schemas/results.ResultStatus'
        result:
          description: Real and imaginary components of complex estimation result value<br>Provided only for successful results
          type: array
          items:
            type: number
          minItems: 1
          maxItems: 2
          example:
            - 1.5
            - 2.8
        reason:
          description: Provided only for unsuccessful (failed, cancelled) results
          type: string
        transpiledCode:
          description: Provided only for tasks with transpilation (tasks created with skipTranspilation=False), if traspilation was successfull
          type: string
        qubitAllocation:
          description: Parameter valid only for QPU devices
          type: object
          additionalProperties:
            type: integer
          example:
            '0': 0
            '1': 4
      required:
        - taskId
        - status