# backend/oqtopus_cloud/provider/schemas/error.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2024-08-29T09:11:46+00:00
#   version:   0.25.8

from __future__ import annotations

from pydantic import BaseModel


class BadRequest(BaseModel):
    detail: str


class NotFoundError(BaseModel):
    detail: str


class InternalServerError(BaseModel):
    detail: str


class ConflictError(BaseModel):
    detail: str