# backend/oqtopus_cloud/user/schemas/error.py
# generated by datamodel-codegen:
#   filename:  openapi.yaml
#   timestamp: 2024-08-29T08:54:22+00:00
#   version:   0.25.8

from __future__ import annotations

from pydantic import BaseModel


class UnauthorizedError(BaseModel):
    detail: str


class InternalServerError(BaseModel):
    detail: str


class NotFoundError(BaseModel):
    detail: str


class BadRequest(BaseModel):
    detail: str